+++ src/Holding/ViewHolding/HoldingBoundWith/useBoundWithHoldings.js
@@ -1,23 +1,21 @@
-import { useQuery } from 'react-query';
-
-import { useOkapiKy, useNamespace } from '@folio/stripes/core';
+import useChunkedCQLFetch from '../../../hooks/useChunkedCQLFetch';
 
 const useBoundWithHoldings = (boundWithItems) => {
-  const ky = useOkapiKy();
-  const [namespace] = useNamespace({ key: 'boundWithHoldings' });
+  const holdingsRecordIds = boundWithItems?.map(x => x.holdingsRecordId);
 
-  const holdingRecordIds = boundWithItems?.map(x => x.holdingsRecordId);
-  const queryIds = holdingRecordIds.join(' or ');
-
-  const { data, isLoading } = useQuery(
-    [namespace, queryIds],
-    () => ky.get(`holdings-storage/holdings?query=id=(${queryIds})`).json(),
-    { enabled: Boolean(queryIds) }
-  );
+  const { items: holdingsRecords, isLoading } = useChunkedCQLFetch({
+    ids: holdingsRecordIds,
+    endpoint: 'holdings-storage/holdings',
+    reduceFunction: (holdingQueries) => (
+      holdingQueries.reduce((acc, curr) => {
+        return [...acc, ...(curr?.data?.holdingsRecords ?? [])];
+      }, [])
+    )
+  });
 
   return {
     isLoading,
-    boundWithHoldings: data?.holdingsRecords || [],
+    boundWithHoldings: holdingsRecords,
   };
 };